feat(sdk): add ErrorLog, clean animation
